# openapi.yaml
openapi: 3.0.3
info:
  title: ZTN_SIM IAMaaS API
  version: "1.0.0"
  description: |
    Versioned REST contract for IAMaaS tenants and core platform operations. The schema documents
    authentication, token issuance, trust-score inspection, policy management, audit/log retrieval,
    and supporting wallet/transaction flows. It is intended to be consumed by the auth-gateway,
    policy decision point, external HMS integrations, and telecom or mobile-money pilots.
servers:
  - url: https://localhost.localdomain.com
    description: Local HTTPS (Podman + nginx)
  - url: http://localhost.localdomain.com
    description: Local HTTP (if configured)
security:
  - BearerAuth: []
    ApiKeyAuth: []
paths:
...
  /transactions:
    post:
      tags: [Transactions]
      summary: Create a transfer or withdrawal with TOTP verification
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreateRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
    get:
      tags: [Transactions]
      summary: List transactions for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
  /transactions/{transactionId}:
    put:
      tags: [Transactions]
      summary: Update transaction status/details
      security:
        - BearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdateRequest'
      responses:
        '200': { $ref: '#/components/responses/MessageResponse' }
    delete:
      tags: [Transactions]
      summary: Cancel a transaction
      security:
        - BearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204': { description: Transaction cancelled }
  /user/initiated-withdrawal:
    post:
      tags: [Transactions]
      summary: Begin a withdrawal awaiting agent approval
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawalRequest'
      responses:
        '201': { $ref: '#/components/responses/MessageResponse' }
  /verify-transaction-otp:
    post:
      tags: [Transactions]
      summary: Verify OTP for sensitive transactions
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionOtpVerifyRequest'
      responses:
        '200': { $ref: '#/components/responses/MessageResponse' }
...
  /agent/transaction:
    post:
      tags: [Agents]
      summary: Process agent transaction
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentTransactionRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
  /agent/transactions:
    get:
      tags: [Agents]
      summary: List transactions managed by agent
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
...
    MessageResponse:
      description: Operation succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
  schemas:
...
    Message:
      type: object
      properties:
        message:
          type: string
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
          format: float
        status:
          type: string
        type:
          type: string
    TransactionCreateRequest:
      type: object
      required: [amount, type]
      properties:
        amount:
          type: number
          format: float
        type:
          type: string
          description: transfer or withdrawal
        totp_code:
          type: string
    TransactionUpdateRequest:
      type: object
      properties:
        status:
          type: string
        note:
          type: string
    WithdrawalRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          format: float
        destination:
          type: string
    TransactionOtpVerifyRequest:
      type: object
      required: [transaction_id, otp]
      properties:
        transaction_id:
          type: integer
        otp:
          type: string
...
    AgentTransactionRequest:
      type: object
      properties:
        amount:
          type: number
          format: float
        user_id:
          type: integer
...